Added new mode, changed font for numbers
Added a mode that requires the user to convert from Kaktovik numerals to Arabic numerals.

Changed the InupiaqNumbers font to have 0 as the first character instead of the last character. This allows for a simpler way of representing text, as there is no longer a need to make an exception for zero. Future fonts should set the Kaktovik numerals to zero at unicode E5A0 (or 58784 in decimal). The KN macro can also be used to add this number to the Kaktovik numeral that is to be displayed.