Completed Numbers to Words
The numbers to words mode has been completed. This should be another stopping point until major reorganization can occur, as the modules are becoming more interconnected in ways that the previous method is no longer scaleable.

Fixed a bug where Inunnaq had a space in the macro, and agliaq could be followed by agliaq.

Fixed misspellings of Aarigaa.